requirements.txt
asgiref>=3.8.1
brotli>=1.1.0
certifi>=2024.6.2
charset-normalizer>=3.3.2
django>=5.0.6
django-cors-headers>=4.4.0
django-phonenumber-field>=8.0.0
djangorestframework>=3.15.2
idna>=3.7
mutagen>=1.47.0
phonenumbers>=8.13.39
pillow>=10.3.0
pycryptodomex>=3.20.0
requests>=2.32.3
sqlparse>=0.5.0
urllib3>=2.2.2
websockets>=12.0
yt-dlp>=2024.7.9
mailersend>=0.5.0
django-environ
sib_api_v3_sdk
cloudinary
psycopg2-binary